Feat: Create dedicated weather forecast page
This commit introduces a new, dedicated weather page to display the 10-day forecast for Rochester, MN. This replaces the previous `TravelStrip` component, which had data accuracy and mobile layout issues.

Key changes:
- Created a new "Weather" page at `/weather`.
- Created a new `Forecast` component to display the weather data.
- Added a "Weather" link to the main navigation bar.
- Corrected the latitude and longitude for Rochester, MN, in the API call to ensure accurate data.
- The new forecast page is fully responsive and provides a clean, user-friendly layout on all devices.
- Removed the old `TravelStrip.tsx` component and its associated test file.
